First cut at a big scrollable table using Calm^2.
Thanks to @justinwoo for the idea!

This version is different in two important ways:

1. When window width changes, styles are updated.
2. Only changes are rendered.

Regarding 2, you can see that by noticing that the random numbers do not
change on visible rows as you scroll.  You can also see that by
inspecting the table element and looking at what gets changed when you
scroll.  This change was done intentionally.  It just as easy to compute
the random numbers after each scrollTop change.

Note that this big table is also fully reactive (though I haven't yet
tested the behavior).  The model can be made into a observable property,
so that one can change the table parameters (everything in the `mock`
object) and the component will immediately adjust.